requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Gunicorn==0.1.1
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1